backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
httpx==0.27.0
starlette==0.36.3
python-multipart==0.0.9
pytest==8.3.4
pytest-asyncio==0.24.0
anyio==4.8.0
duckdb>=0.10.0
psycopg2-binary>=2.9.9
